functions/requirements.txt
firebase_functions~=0.1
injector~=0.22
shortuuid~=1.0
requests~=2.32
pymupdf~=1.26
python-dotenv~=1.1